requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
whitenoise==6.3.0
django-grappelli==3.0.4
psycopg2-binary==2.9.5
dj-database-url==1.2.0
django-on-heroku==1.1.2
gunicorn==20.1.0
django-report-builder==6.4.2
django-crum==0.7.9
django-simple-history==3.2.0
wagtail==4.2
pylint==2.16.2
python-liquid==1.9.1